JWT Manager + additional_claims (ONLY BACKEND)

+++ backend/src/Blueprints/task_app.py
@@ -2,8 +2,11 @@ from flask import Blueprint, make_response, jsonify, request
 from uuid import uuid1
 from models.task_app import Tasks
 from repository.task_app import TasksRepository
+from flask_jwt_extended import jwt_required, get_jwt, get_jwt_identity, JWTManager
+
 
 tasks = Blueprint('tasks', __name__, url_prefix="/tasks")
+jwtManager = JWTManager()
 task_repository = TasksRepository()
 
 @tasks.get('/')
@@ -25,34 +28,40 @@ def get_one_task(id):
         return make_response(e.__str__(), 400)
     
 @tasks.delete('/delete/<id>')
+@jwt_required()
 def delete_by_id(id):
-    try:
-        task_repository.delete(id)
-        return make_response("Deleted")
-    except Exception as e:
-        return make_response(e.__str__(), 400)
+    claims= get_jwt()
+    if claims["isAdmin"]:
+        try:
+            task_repository.delete(id)
+            return make_response("Deleted")
+        except Exception as e:
+            return make_response(e.__str__(), 400)
 
 @tasks.post('/create')
+@jwt_required()
 def create():
-    try:
-        task_data = request.get_json()
-        #Create data variables from request
-        task_id = str(uuid1())
-        employee_id = task_data["employee_id"]
-        assigned_by_id = task_data["assigned_by_id"]
-        company_id = task_data["company_id"]
-        title = task_data["title"]
-        description = task_data["description"]
-        due_date = task_data["due_date"] 
-        is_completed = task_data["is_completed"]
-        #Create task object from model
-        created_task = Tasks(task_id, employee_id, assigned_by_id, company_id, title,description,due_date,is_completed)
-        # Add to db
-        task_repository.create(created_task)
-        
-        return make_response("task is created", 200)
-    except Exception as e:
-        return make_response(e.__str__(), 400)
+    claims= get_jwt()
+    if claims["isAdmin"]:
+        try:
+            task_data = request.get_json()
+            #Create data variables from request
+            task_id = str(uuid1())
+            employee_id = task_data["employee_id"]
+            assigned_by_id = task_data["assigned_by_id"]
+            company_id = task_data["company_id"]
+            title = task_data["title"]
+            description = task_data["description"]
+            due_date = task_data["due_date"] 
+            is_completed = task_data["is_completed"]
+            #Create task object from model
+            created_task = Tasks(task_id, employee_id, assigned_by_id, company_id, title,description,due_date,is_completed)
+            # Add to db
+            task_repository.create(created_task)
+            
+            return make_response("task is created", 200)
+        except Exception as e:
+            return make_response(e.__str__(), 400)
     
 @tasks.put('/done/≤id>')
 def mark_as_done():
